pyproject.toml
[project]
name = "glacier-watch-ingest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "geoalchemy2>=0.18.1",
    "geopandas>=1.1.1",
    "psycopg>=3.3.2",
    "pystac-client>=0.9.0",
    "python-dotenv>=1.2.1",
    "python-json-logger>=4.0.0",
    "pyyaml>=6.0.3",
    "rasterio>=1.4.3",
    "requests>=2.32.5",
    "rioxarray>=0.20.0",
    "shapely>=2.1.2",
    "sqlmodel>=0.0.27",
    "tqdm>=4.67.1",
    "xarray>=2025.12.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "ruff>=0.14.8",
]
